APP LISTのスロットサイズ可変機能とウィジェットリサイズのライブスナップを追加
- APP LISTのアイコン密度をS/M/Lで切り替えるヘッダーボタンを追加
- ウィジェットのリサイズ・移動時、ドラッグ中に指を離す前からスナップ先が
  リアルタイムに見えるようにした
- PlacedWidgetのcolSpan/rowSpanをFloat化し、APP LISTのリサイズはアイコン1個分の
  固定サイズを単位に1行・1列ずつスナップするようにした
- APP LISTの列数・行数を、ウィジェットの現在の外枠サイズ（リサイズ中のライブ
  プレビュー値を含む）から動的に算出するようにし、リサイズ中の表示崩れを解消
- EditHandleのpointerInputが初回描画時の古いクロージャに固定され、
  2回目以降のドラッグ確定時に元のサイズ・位置へ戻ってしまう不具合を
  rememberUpdatedStateで修正

+++ app/src/main/java/com/example/girdlauncher/model/PlacedWidget.kt
@@ -6,15 +6,16 @@ package com.example.girdlauncher.model
  * @property type ウィジェットの種類。
  * @property col 配置されている列（0始まり）。
  * @property row 配置されている行（0始まり）。
- * @property colSpan 横方向に占めるセル数。
- * @property rowSpan 縦方向に占めるセル数。
+ * @property colSpan 横方向に占めるセル数。ウィジェットの種類によっては、アイコンサイズ単位での
+ *   細かいリサイズに対応するため小数値を取り得る。
+ * @property rowSpan 縦方向に占めるセル数。[colSpan]と同様に小数値を取り得る。
  */
 data class PlacedWidget(
     val type: WidgetPanel,
     val col: Int,
     val row: Int,
-    val colSpan: Int,
-    val rowSpan: Int
+    val colSpan: Float,
+    val rowSpan: Float
 ) {
     /** [other]と矩形が重なっているかどうか。 */
     fun overlaps(other: PlacedWidget): Boolean {
